+++ R/runSimulation.R
@@ -89,12 +89,8 @@ runSimulation <- function(name,
     assertthat::assert_that(file.exists(path_solver))
     assertthat::assert_that(mode %in% c("economy", "adequacy", "draft", "expansion"))
     
-    
     ##Test version of antares solver
-    solver <- unlist(gsub("-solver.exe", "", path_solver))
-    solver <- strsplit(solver, "antares-")[[1]]
-    solver <- solver[[length(solver)]]
-    version_solver <- substr(solver, 1, 1)
+    version_solver <- .get_version_solver_from_path_solver(path_solver = path_solver)
     version_study <- substr(opts$antaresVersion,1,1)
     
     if (version_solver != version_study) {