+++ NEWS.md
@@ -3,11 +3,14 @@
 
 # antaresEditObject 0.9.4.9000
 
-
 NEW FEATURES :
 
 * `createDistrict()` uses a specific endpoint to create the district
 * `editDistrict()` allows the user to edit a district. Use of a specific endpoint in API mode.
+
+BUGFIXES :    
+
+* `.get_version_solver_from_path_solver()` get solver version by running command line instead of searching in the path.
 
 OPTIMIZATION :